requirements.txt
# Core dependencies for JobHelper

# Vector embeddings and similarity
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Web scraping (for future use)
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data handling
pandas>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Document parsing
markdown>=3.5.0

# Resume generation
jinja2>=3.1.0

# Web interface
flask>=3.0.0
flask-cors>=4.0.0

# AI-powered job scraping and analysis
openai>=1.0.0
